+++ lib/widgets/home/navbar/new_transaction_button.dart
@@ -2,6 +2,7 @@ import 'package:flow/entity/transaction.dart';
 import 'package:flow/l10n/extensions.dart';
 import 'package:flow/l10n/named_enum.dart';
 import 'package:flow/main.dart';
+import 'package:flow/prefs.dart';
 import 'package:flow/theme/navbar_theme.dart';
 import 'package:flow/theme/theme.dart';
 import 'package:flutter/material.dart' hide Flow;
@@ -24,80 +25,65 @@ class _NewTransactionButtonState extends State<NewTransactionButton> {
   Widget build(BuildContext context) {
     final NavbarTheme navbarTheme = Theme.of(context).extension<NavbarTheme>()!;
 
-    return PieMenu(
-      theme: Flow.of(context).pieTheme.copyWith(
-            customAngle: 90.0,
-            customAngleDiff: 48.0,
-            radius: 108.0,
-            customAngleAnchor: PieAnchor.center,
-            leftClickShowsMenu: true,
-            rightClickShowsMenu: true,
-            delayDuration: Duration.zero,
-          ),
-      onToggle: onToggle,
-      actions: [
-        PieAction(
-          tooltip: Text(TransactionType.transfer.localizedNameContext(context)),
-          onSelect: () => widget.onActionTap(TransactionType.transfer),
-          child: const Icon(
-            Symbols.compare_arrows_rounded,
-            weight: 800.0,
-          ),
-          buttonTheme: PieButtonTheme(
-            backgroundColor: context.colorScheme.secondary,
-            iconColor: context.colorScheme.onSecondary,
-          ),
-        ),
-        PieAction(
-          tooltip: Text(TransactionType.income.localizedNameContext(context)),
-          onSelect: () => widget.onActionTap(TransactionType.income),
-          child: const Icon(
-            Symbols.stat_2_rounded,
-            weight: 800.0,
-          ),
-          buttonTheme: PieButtonTheme(
-            backgroundColor: context.flowColors.income,
-            iconColor: context.colorScheme.onError,
-          ),
-        ),
-        PieAction(
-          tooltip: Text(TransactionType.expense.localizedNameContext(context)),
-          onSelect: () => widget.onActionTap(TransactionType.expense),
-          child: const Icon(
-            Symbols.stat_minus_2_rounded,
-            weight: 800.0,
-          ),
-          buttonTheme: PieButtonTheme(
-            backgroundColor: context.flowColors.expense,
-            iconColor: context.colorScheme.onError,
-          ),
-        ),
-      ],
-      child: StatefulBuilder(
-        builder: (context, setState) => Tooltip(
-          message: "transaction.new".t(context),
-          child: Material(
-            color: navbarTheme.transactionButtonBackgroundColor,
-            shape: RoundedRectangleBorder(
-              borderRadius: BorderRadius.circular(32.0),
-            ),
-            child: Padding(
-              padding: const EdgeInsets.all(20.0),
-              child: AnimatedRotation(
-                turns: _buttonRotationTurns,
-                duration: const Duration(milliseconds: 600),
-                child: Icon(
-                  Symbols.add_rounded,
-                  fill: 0.0,
-                  color: navbarTheme.transactionButtonForegroundColor,
-                  weight: 600.0,
+    return ValueListenableBuilder(
+        valueListenable:
+            LocalPreferences().transactionButtonOrder.valueNotifier,
+        builder: (context, buttonOrder, child) {
+          buttonOrder ??= TransactionType.values;
+
+          return PieMenu(
+            theme: Flow.of(context).pieTheme.copyWith(
+                  customAngle: 90.0,
+                  customAngleDiff: 48.0,
+                  radius: 108.0,
+                  customAngleAnchor: PieAnchor.center,
+                  leftClickShowsMenu: true,
+                  rightClickShowsMenu: true,
+                  delayDuration: Duration.zero,
+                ),
+            onToggle: onToggle,
+            actions: [
+              for (final transactionType in buttonOrder)
+                PieAction(
+                  tooltip: Text(transactionType.localizedNameContext(context)),
+                  onSelect: () => widget.onActionTap(transactionType),
+                  child: Icon(
+                    transactionType.icon,
+                    weight: 800.0,
+                  ),
+                  buttonTheme: PieButtonTheme(
+                    backgroundColor:
+                        transactionType.actionBackgroundColor(context),
+                    iconColor: transactionType.actionColor(context),
+                  ),
+                ),
+            ],
+            child: StatefulBuilder(
+              builder: (context, setState) => Tooltip(
+                message: "transaction.new".t(context),
+                child: Material(
+                  color: navbarTheme.transactionButtonBackgroundColor,
+                  shape: RoundedRectangleBorder(
+                    borderRadius: BorderRadius.circular(32.0),
+                  ),
+                  child: Padding(
+                    padding: const EdgeInsets.all(20.0),
+                    child: AnimatedRotation(
+                      turns: _buttonRotationTurns,
+                      duration: const Duration(milliseconds: 600),
+                      child: Icon(
+                        Symbols.add_rounded,
+                        fill: 0.0,
+                        color: navbarTheme.transactionButtonForegroundColor,
+                        weight: 600.0,
+                      ),
+                    ),
+                  ),
                 ),
               ),
             ),
-          ),
-        ),
-      ),
-    );
+          );
+        });
   }
 
   void onToggle(bool toggled) {